+++ include/ceed/backend.h
@@ -32,6 +32,39 @@
 #define CEED_COMPOSITE_MAX 16
 #define CEED_EPSILON 1E-16
 
+/**
+  @ingroup Ceed
+  This macro provides the ablitiy to disable optimization flags for functions that
+  are sensitive to floting point optimizations.
+**/
+#ifndef CeedPragmaOptimizeOff
+#  if defined(__clang__)
+#    define CeedPragmaOptimizeOff _Pragma("clang optimize off")
+#  elif defined(__GNUC__)
+#    define CeedPragmaOptimizeOff _Pragma("GCC push_options") _Pragma("GCC optimize 0")
+#  elif defined(__INTEL_COMPILER) || defined(__INTEL_LLVM_COMPILER)
+#    define CeedPragmaOptimizeOff _Pragma("optimize('', off)")
+#  else
+#    define CeedPragmaOptimizeOff
+#  endif
+#endif
+
+/**
+  @ingroup Ceed
+  This macro restores previously set optimization flags after CeedPragmaOptimizeOff.
+**/
+#ifndef CeedPragmaOptimizeOn
+#  if defined(__clang__)
+#    define CeedPragmaOptimizeOn _Pragma("clang optimize on")
+#  elif defined(__GNUC__)
+#    define CeedPragmaOptimizeOn _Pragma("GCC pop_options")
+#  elif defined(__INTEL_COMPILER) || defined(__INTEL_LLVM_COMPILER)
+#    define CeedPragmaOptimizeOff _Pragma("optimize('', on)")
+#  else
+#    define CeedPragmaOptimizeOn
+#  endif
+#endif
+
 /// CEED_DEBUG_COLOR default value, forward CeedDebug* declarations & macros
 #ifndef CEED_DEBUG_COLOR
 #define CEED_DEBUG_COLOR 0