+++ interface/ceed-basis.c
@@ -83,7 +83,7 @@ static int CeedHouseholderReflect(CeedScalar *A, const CeedScalar *v,
   @param[in,out] A  Matrix to apply Householder Q to, in place
   @param Q          Householder Q matrix
   @param tau        Householder scaling factors
-  @param t_mode    Transpose mode for application
+  @param t_mode     Transpose mode for application
   @param m          Number of rows in A
   @param n          Number of columns in A
   @param k          Number of elementary reflectors in Q, k<m
@@ -229,7 +229,7 @@ int CeedBasisGetCollocatedGrad(CeedBasis basis, CeedScalar *collo_grad_1d) {
       collo_grad_1d[j+Q_1d*i] = 0;
   }
 
-  // Apply Qtranspose, collograd = collo_grad Q_transpose
+  // Apply Qtranspose, collo_grad = collo_grad Q_transpose
   ierr = CeedHouseholderApplyQ(collo_grad_1d, interp_1d, tau, CEED_NOTRANSPOSE,
                                Q_1d, Q_1d, P_1d, 1, Q_1d); CeedChk(ierr);
 
@@ -895,7 +895,7 @@ int CeedBasisGetNumQuadraturePoints(CeedBasis basis, CeedInt *Q) {
 /**
   @brief Get total number of quadrature points (in 1 dimension) of a CeedBasis
 
-  @param basis     CeedBasis
+  @param basis      CeedBasis
   @param[out] Q_1d  Variable to store number of quadrature points
 
   @return An error code: 0 - success, otherwise - failure
@@ -933,7 +933,7 @@ int CeedBasisGetQRef(CeedBasis basis, const CeedScalar **q_ref) {
   @brief Get quadrature weights of quadrature points (in dim dimensions)
          of a CeedBasis
 
-  @param basis         CeedBasis
+  @param basis          CeedBasis
   @param[out] q_weight  Variable to store quadrature weights
 
   @return An error code: 0 - success, otherwise - failure
@@ -1093,8 +1093,8 @@ int CeedBasisDestroy(CeedBasis *basis) {
 /**
   @brief Construct a Gauss-Legendre quadrature
 
-  @param Q               Number of quadrature points (integrates polynomials of
-                           degree 2*Q-1 exactly)
+  @param Q                 Number of quadrature points (integrates polynomials of
+                             degree 2*Q-1 exactly)
   @param[out] q_ref_1d     Array of length Q to hold the abscissa on [-1, 1]
   @param[out] q_weight_1d  Array of length Q to hold the weights
 
@@ -1147,8 +1147,8 @@ int CeedGaussQuadrature(CeedInt Q, CeedScalar *q_ref_1d,
 /**
   @brief Construct a Gauss-Legendre-Lobatto quadrature
 
-  @param Q               Number of quadrature points (integrates polynomials of
-                           degree 2*Q-3 exactly)
+  @param Q                 Number of quadrature points (integrates polynomials of
+                             degree 2*Q-3 exactly)
   @param[out] q_ref_1d     Array of length Q to hold the abscissa on [-1, 1]
   @param[out] q_weight_1d  Array of length Q to hold the weights
 
@@ -1241,6 +1241,10 @@ int CeedQRFactorization(Ceed ceed, CeedScalar *mat, CeedScalar *tau,
   // LCOV_EXCL_STOP
 
   for (CeedInt i=0; i<n; i++) {
+    if (i >= m-1) { // last row of matrix, no reflection needed
+      tau[i] = 0.;
+      break;
+    }
     // Calculate Householder vector, magnitude
     CeedScalar sigma = 0.0;
     v[i] = mat[i+n*i];
@@ -1249,20 +1253,19 @@ int CeedQRFactorization(Ceed ceed, CeedScalar *mat, CeedScalar *tau,
       sigma += v[j] * v[j];
     }
     CeedScalar norm = sqrt(v[i]*v[i] + sigma); // norm of v[i:m]
-    CeedScalar Rii = -copysign(norm, v[i]);
-    v[i] -= Rii;
+    CeedScalar R_ii = -copysign(norm, v[i]);
+    v[i] -= R_ii;
     // norm of v[i:m] after modification above and scaling below
     //   norm = sqrt(v[i]*v[i] + sigma) / v[i];
     //   tau = 2 / (norm*norm)
     tau[i] = 2 * v[i]*v[i] / (v[i]*v[i] + sigma);
-
     for (CeedInt j=i+1; j<m; j++)
       v[j] /= v[i];
 
     // Apply Householder reflector to lower right panel
     CeedHouseholderReflect(&mat[i*n+i+1], &v[i], tau[i], m-i, n-i-1, n, 1);
     // Save v
-    mat[i+n*i] = Rii;
+    mat[i+n*i] = R_ii;
     for (CeedInt j=i+1; j<m; j++)
       mat[i+n*j] = v[j];
   }
@@ -1282,6 +1285,7 @@ int CeedQRFactorization(Ceed ceed, CeedScalar *mat, CeedScalar *tau,
 
   @ref Utility
 **/
+CeedPragmaOptimizeOff
 int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
                                     CeedScalar *lambda, CeedInt n) {
   // Check bounds for clang-tidy
@@ -1291,10 +1295,10 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
                      "Cannot compute symmetric Schur decomposition of scalars");
   // LCOV_EXCL_STOP
 
-  CeedScalar v[n-1], tau[n-1], matT[n*n];
+  CeedScalar v[n-1], tau[n-1], mat_T[n*n];
 
-  // Copy mat to matT and set mat to I
-  memcpy(matT, mat, n*n*sizeof(mat[0]));
+  // Copy mat to mat_T and set mat to I
+  memcpy(mat_T, mat, n*n*sizeof(mat[0]));
   for (CeedInt i=0; i<n; i++)
     for (CeedInt j=0; j<n; j++)
       mat[j+n*i] = (i==j) ? 1 : 0;
@@ -1303,67 +1307,66 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
   for (CeedInt i=0; i<n-1; i++) {
     // Calculate Householder vector, magnitude
     CeedScalar sigma = 0.0;
-    v[i] = matT[i+n*(i+1)];
+    v[i] = mat_T[i+n*(i+1)];
     for (CeedInt j=i+1; j<n-1; j++) {
-      v[j] = matT[i+n*(j+1)];
+      v[j] = mat_T[i+n*(j+1)];
       sigma += v[j] * v[j];
     }
     CeedScalar norm = sqrt(v[i]*v[i] + sigma); // norm of v[i:n-1]
-    CeedScalar Rii = -copysign(norm, v[i]);
-    v[i] -= Rii;
+    CeedScalar R_ii = -copysign(norm, v[i]);
+    v[i] -= R_ii;
     // norm of v[i:m] after modification above and scaling below
     //   norm = sqrt(v[i]*v[i] + sigma) / v[i];
     //   tau = 2 / (norm*norm)
-    if (sigma > 10*CEED_EPSILON)
-      tau[i] = 2 * v[i]*v[i] / (v[i]*v[i] + sigma);
-    else
-      tau[i] = 0;
-
+    tau[i] = 2 * v[i]*v[i] / (v[i]*v[i] + sigma);
     for (CeedInt j=i+1; j<n-1; j++)
       v[j] /= v[i];
 
     // Update sub and super diagonal
-    matT[i+n*(i+1)] = Rii;
-    matT[(i+1)+n*i] = Rii;
     for (CeedInt j=i+2; j<n; j++) {
-      matT[i+n*j] = 0; matT[j+n*i] = 0;
+      mat_T[i+n*j] = 0; mat_T[j+n*i] = 0;
     }
     // Apply symmetric Householder reflector to lower right panel
-    CeedHouseholderReflect(&matT[(i+1)+n*(i+1)], &v[i], tau[i],
+    CeedHouseholderReflect(&mat_T[(i+1)+n*(i+1)], &v[i], tau[i],
                            n-(i+1), n-(i+1), n, 1);
-    CeedHouseholderReflect(&matT[(i+1)+n*(i+1)], &v[i], tau[i],
+    CeedHouseholderReflect(&mat_T[(i+1)+n*(i+1)], &v[i], tau[i],
                            n-(i+1), n-(i+1), 1, n);
+
     // Save v
+    mat_T[i+n*(i+1)] = R_ii;
+    mat_T[(i+1)+n*i] = R_ii;
     for (CeedInt j=i+1; j<n-1; j++) {
-      matT[i+n*(j+1)] = v[j];
+      mat_T[i+n*(j+1)] = v[j];
     }
   }
   // Backwards accumulation of Q
   for (CeedInt i=n-2; i>=0; i--) {
-    v[i] = 1;
-    for (CeedInt j=i+1; j<n-1; j++) {
-      v[j] = matT[i+n*(j+1)];
-      matT[i+n*(j+1)] = 0;
+    if (tau[i] > 0.0) {
+      v[i] = 1;
+      for (CeedInt j=i+1; j<n-1; j++) {
+        v[j] = mat_T[i+n*(j+1)];
+        mat_T[i+n*(j+1)] = 0;
+      }
+      CeedHouseholderReflect(&mat[(i+1)+n*(i+1)], &v[i], tau[i],
+                             n-(i+1), n-(i+1), n, 1);
     }
-    CeedHouseholderReflect(&mat[(i+1)+n*(i+1)], &v[i], tau[i],
-                           n-(i+1), n-(i+1), n, 1);
   }
 
   // Reduce sub and super diagonal
-  CeedInt p = 0, q = 0, itr = 0, maxitr = n*n*n;
-  CeedScalar tol = 10*CEED_EPSILON;
+  CeedInt p = 0, q = 0, itr = 0, max_itr = n*n*n*n;
+  CeedScalar tol = CEED_EPSILON;
 
-  while (q < n && itr < maxitr) {
+  while (itr < max_itr) {
     // Update p, q, size of reduced portions of diagonal
     p = 0; q = 0;
     for (CeedInt i=n-2; i>=0; i--) {
-      if (fabs(matT[i+n*(i+1)]) < tol)
+      if (fabs(mat_T[i+n*(i+1)]) < tol)
         q += 1;
       else
         break;
     }
-    for (CeedInt i=0; i<n-1-q; i++) {
-      if (fabs(matT[i+n*(i+1)]) < tol)
+    for (CeedInt i=0; i<n-q-1; i++) {
+      if (fabs(mat_T[i+n*(i+1)]) < tol)
         p += 1;
       else
         break;
@@ -1371,14 +1374,14 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
     if (q == n-1) break; // Finished reducing
 
     // Reduce tridiagonal portion
-    CeedScalar tnn = matT[(n-1-q)+n*(n-1-q)],
-               tnnm1 = matT[(n-2-q)+n*(n-1-q)];
-    CeedScalar d = (matT[(n-2-q)+n*(n-2-q)] - tnn)/2;
-    CeedScalar mu = tnn - tnnm1*tnnm1 /
-                    (d + copysign(sqrt(d*d + tnnm1*tnnm1), d));
-    CeedScalar x = matT[p+n*p] - mu;
-    CeedScalar z = matT[p+n*(p+1)];
-    for (CeedInt k=p; k<n-1-q; k++) {
+    CeedScalar t_nn = mat_T[(n-1-q)+n*(n-1-q)],
+               t_nnm1 = mat_T[(n-2-q)+n*(n-1-q)];
+    CeedScalar d = (mat_T[(n-2-q)+n*(n-2-q)] - t_nn)/2;
+    CeedScalar mu = t_nn - t_nnm1*t_nnm1 /
+                    (d + copysign(sqrt(d*d + t_nnm1*t_nnm1), d));
+    CeedScalar x = mat_T[p+n*p] - mu;
+    CeedScalar z = mat_T[p+n*(p+1)];
+    for (CeedInt k=p; k<n-q-1; k++) {
       // Compute Givens rotation
       CeedScalar c = 1, s = 0;
       if (fabs(z) > tol) {
@@ -1392,32 +1395,34 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
       }
 
       // Apply Givens rotation to T
-      CeedGivensRotation(matT, c, s, CEED_NOTRANSPOSE, k, k+1, n, n);
-      CeedGivensRotation(matT, c, s, CEED_TRANSPOSE, k, k+1, n, n);
+      CeedGivensRotation(mat_T, c, s, CEED_NOTRANSPOSE, k, k+1, n, n);
+      CeedGivensRotation(mat_T, c, s, CEED_TRANSPOSE, k, k+1, n, n);
 
       // Apply Givens rotation to Q
       CeedGivensRotation(mat, c, s, CEED_NOTRANSPOSE, k, k+1, n, n);
 
       // Update x, z
       if (k < n-q-2) {
-        x = matT[k+n*(k+1)];
-        z = matT[k+n*(k+2)];
+        x = mat_T[k+n*(k+1)];
+        z = mat_T[k+n*(k+2)];
       }
     }
     itr++;
   }
+
   // Save eigenvalues
   for (CeedInt i=0; i<n; i++)
-    lambda[i] = matT[i+n*i];
+    lambda[i] = mat_T[i+n*i];
 
   // Check convergence
-  if (itr == maxitr && q < n-1)
+  if (itr == max_itr && q < n-1)
     // LCOV_EXCL_START
     return CeedError(ceed, CEED_ERROR_MINOR,
                      "Symmetric QR failed to converge");
   // LCOV_EXCL_STOP
   return CEED_ERROR_SUCCESS;
 }
+CeedPragmaOptimizeOn
 
 /**
   @brief Return Simultaneous Diagonalization of two matrices. This solves the
@@ -1427,9 +1432,9 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
            is equivalent to the LAPACK routine 'sygv' with TYPE = 1.
 
   @param ceed         A Ceed context for error handling
-  @param[in] mat_A     Row-major matrix to be factorized with eigenvalues
-  @param[in] mat_B     Row-major matrix to be factorized to identity
-  @param[out] x       Row-major orthogonal matrix
+  @param[in] mat_A    Row-major matrix to be factorized with eigenvalues
+  @param[in] mat_B    Row-major matrix to be factorized to identity
+  @param[out] mat_X   Row-major orthogonal matrix
   @param[out] lambda  Vector of length n of generalized eigenvalues
   @param n            Number of rows/columns
 
@@ -1437,50 +1442,80 @@ int CeedSymmetricSchurDecomposition(Ceed ceed, CeedScalar *mat,
 
   @ref Utility
 **/
+CeedPragmaOptimizeOff
 int CeedSimultaneousDiagonalization(Ceed ceed, CeedScalar *mat_A,
-                                    CeedScalar *mat_B, CeedScalar *x,
+                                    CeedScalar *mat_B, CeedScalar *mat_X,
                                     CeedScalar *lambda, CeedInt n) {
   int ierr;
-  CeedScalar *mat_C, *mat_G, *vecD;
+  CeedScalar *mat_C, *mat_G, *vec_D;
   ierr = CeedCalloc(n*n, &mat_C); CeedChk(ierr);
   ierr = CeedCalloc(n*n, &mat_G); CeedChk(ierr);
-  ierr = CeedCalloc(n, &vecD); CeedChk(ierr);
+  ierr = CeedCalloc(n, &vec_D); CeedChk(ierr);
 
   // Compute B = G D G^T
   memcpy(mat_G, mat_B, n*n*sizeof(mat_B[0]));
-  ierr = CeedSymmetricSchurDecomposition(ceed, mat_G, vecD, n); CeedChk(ierr);
-  for (CeedInt i=0; i<n; i++)
-    vecD[i] = sqrt(vecD[i]);
+  ierr = CeedSymmetricSchurDecomposition(ceed, mat_G, vec_D, n); CeedChk(ierr);
+
+  // Sort eigenvalues
+  for (CeedInt i=n-1; i>=0; i--)
+    for (CeedInt j=0; j<i; j++) {
+      if (fabs(vec_D[j]) > fabs(vec_D[j+1])) {
+        CeedScalar temp;
+        temp = vec_D[j]; vec_D[j] = vec_D[j+1]; vec_D[j+1] = temp;
+        for (CeedInt k=0; k<n; k++) {
+          temp = mat_G[k*n+j]; mat_G[k*n+j] = mat_G[k*n+j+1]; mat_G[k*n+j+1] = temp;
+        }
+      }
+    }
 
   // Compute C = (G D^1/2)^-1 A (G D^1/2)^-T
   //           = D^-1/2 G^T A G D^-1/2
+  // -- D = D^-1/2
   for (CeedInt i=0; i<n; i++)
-    for (CeedInt j=0; j<n; j++)
-      mat_C[j+i*n] = mat_G[i+j*n] / vecD[i];
+    vec_D[i] = 1./sqrt(vec_D[i]);
+  // -- G = G D^-1/2
+  // -- C = D^-1/2 G^T
+  for (CeedInt i=0; i<n; i++)
+    for (CeedInt j=0; j<n; j++) {
+      mat_G[i*n+j] *= vec_D[j];
+      mat_C[j*n+i]  = mat_G[i*n+j];
+    }
+  // -- X = (D^-1/2 G^T) A
   ierr = CeedMatrixMultiply(ceed, (const CeedScalar *)mat_C,
-                            (const CeedScalar *)mat_A, x, n, n, n);
+                            (const CeedScalar *)mat_A, mat_X, n, n, n);
   CeedChk(ierr);
-  for (CeedInt i=0; i<n; i++)
-    for (CeedInt j=0; j<n; j++)
-      mat_G[j+i*n] = mat_G[j+i*n] / vecD[j];
-  ierr = CeedMatrixMultiply(ceed, (const CeedScalar *)x,
+  // -- C = (D^-1/2 G^T A) (G D^-1/2)
+  ierr = CeedMatrixMultiply(ceed, (const CeedScalar *)mat_X,
                             (const CeedScalar *)mat_G, mat_C, n, n, n);
   CeedChk(ierr);
 
   // Compute Q^T C Q = lambda
   ierr = CeedSymmetricSchurDecomposition(ceed, mat_C, lambda, n); CeedChk(ierr);
 
-  // Set x = (G D^1/2)^-T Q
+  // Sort eigenvalues
+  for (CeedInt i=n-1; i>=0; i--)
+    for (CeedInt j=0; j<i; j++) {
+      if (fabs(lambda[j]) > fabs(lambda[j+1])) {
+        CeedScalar temp;
+        temp = lambda[j]; lambda[j] = lambda[j+1]; lambda[j+1] = temp;
+        for (CeedInt k=0; k<n; k++) {
+          temp = mat_C[k*n+j]; mat_C[k*n+j] = mat_C[k*n+j+1]; mat_C[k*n+j+1] = temp;
+        }
+      }
+    }
+
+  // Set X = (G D^1/2)^-T Q
   //       = G D^-1/2 Q
   ierr = CeedMatrixMultiply(ceed, (const CeedScalar *)mat_G,
-                            (const CeedScalar *)mat_C, x, n, n, n);
+                            (const CeedScalar *)mat_C, mat_X, n, n, n);
   CeedChk(ierr);
 
   // Cleanup
   ierr = CeedFree(&mat_C); CeedChk(ierr);
   ierr = CeedFree(&mat_G); CeedChk(ierr);
-  ierr = CeedFree(&vecD); CeedChk(ierr);
+  ierr = CeedFree(&vec_D); CeedChk(ierr);
   return CEED_ERROR_SUCCESS;
 }
+CeedPragmaOptimizeOn
 
 /// @}